Correção para junção dos validators CPF/CNPJ e inserção de validação de números repetidos no campo de Mobile Phone
